+++ vehiculo-service/src/main/java/com/example/vehiculo_service/repository/VehiculoRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface VehiculoRepository extends JpaRepository<Vehiculo, Integer> {
+public interface VehiculoRepository extends JpaRepository<Vehiculo, Long> {
     Optional<Vehiculo> findByPatente(String patente);
 }
